refactor materials and shaders
* consolidated shader code
* exposed msaa as engine config param
