+++ .vscode/launch.json
@@ -8,8 +8,8 @@
         "name": "(lldb) Launch",
         "type": "cppdbg",
         "request": "launch",
-        "program": "${workspaceFolder}/build_debug/samples/sokol/sample_glengine2",
-        "args": ["../resources/models/the_mill/scene.gltf"],
+        "program": "${workspaceFolder}/build_debug/samples/glengine/benchmark_app",
+        "args": [],
         "stopAtEntry": false,
         "cwd": "${workspaceFolder}/build_debug",
         "environment": [],
