+++ glengine/gl_material.h
@@ -7,13 +7,13 @@
 
 namespace glengine {
 
-class ResourceManager;
+class GLEngine;
 
 class Material {
   public:
     virtual ~Material() = default;
 
-    virtual bool init(ResourceManager &rm, sg_primitive_type primitive, sg_index_type = SG_INDEXTYPE_NONE) = 0;
+    virtual bool init(GLEngine &eng, sg_primitive_type primitive, sg_index_type = SG_INDEXTYPE_NONE) = 0;
 
     virtual void update_bindings(sg_bindings &bind) {}
