+++ glengine/gl_material_diffuse.h
@@ -14,11 +14,29 @@ class MaterialDiffuse : public Material {
     : Material() {}
     virtual ~MaterialDiffuse() = default;
 
-    virtual bool init(ResourceManager &rm, sg_primitive_type primitive, sg_index_type idx_type = SG_INDEXTYPE_NONE) override;
+    virtual bool init(GLEngine &eng, sg_primitive_type primitive, sg_index_type idx_type = SG_INDEXTYPE_NONE) override;
 
     virtual void update_bindings(sg_bindings &bind) override;
 
     virtual void apply_uniforms(const common_uniform_params_t &params) override;
+};
+
+class MaterialDiffuseTextured : public Material {
+  public:
+    MaterialDiffuseTextured()
+    : Material() {
+        color = {255, 255, 255, 255};
+    }
+    virtual ~MaterialDiffuseTextured() = default;
+
+    virtual bool init(GLEngine &eng, sg_primitive_type primitive,
+                      sg_index_type idx_type = SG_INDEXTYPE_NONE) override;
+
+    virtual void update_bindings(sg_bindings &bind) override;
+
+    virtual void apply_uniforms(const common_uniform_params_t &params) override;
+
+    sg_image tex_diffuse = {0};
 };
 
 } // namespace glengine
