+++ glengine/gl_material_vertexcolor.h
@@ -14,7 +14,7 @@ class MaterialVertexColor : public Material {
     : Material() {}
     virtual ~MaterialVertexColor() = default;
 
-    virtual bool init(ResourceManager &rm, sg_primitive_type primitive, sg_index_type idx_type = SG_INDEXTYPE_NONE) override;
+    virtual bool init(GLEngine &eng, sg_primitive_type primitive, sg_index_type idx_type = SG_INDEXTYPE_NONE) override;
 
     virtual void update_bindings(sg_bindings &bind) override;
