+++ samples/glengine/sample_gltf.cpp
@@ -4,8 +4,6 @@
 #include "gl_mesh.h"
 #include "gl_prefabs.h"
 #include "gl_material_diffuse.h"
-#include "gl_material_flat.h"
-#include "gl_material_flat_textured.h"
 #include "gl_material_vertexcolor.h"
 #include "gl_material_pbr.h"
 #include "gl_material_pbr_ibl.h"
